+++ tests/Sample/Handlers/Parsed/Response.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rougin\Slytherin\Sample\Handlers;
+namespace Rougin\Slytherin\Sample\Handlers\Parsed;
 
 use Interop\Http\ServerMiddleware\DelegateInterface;
 use Interop\Http\ServerMiddleware\MiddlewareInterface;
@@ -10,14 +10,14 @@ use Psr\Http\Message\ServerRequestInterface;
  * @package Slytherin
  * @author  Rougin Gutib <rougingutib@gmail.com>
  */
-class Parsed implements MiddlewareInterface
+class Response implements MiddlewareInterface
 {
     public function process(ServerRequestInterface $request, DelegateInterface $delegate)
     {
-        $data = array('name' => 'Slytherin');
+        $response = $delegate->process($request);
 
-        $request = $request->withParsedBody($data);
+        $response->getBody()->write('From middleware!');
 
-        return $delegate->process($request);
+        return $response;
     }
 }